# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmniReferral API",
        "description": "REST API for the OmniReferral real estate referral and lead management platform.\n\nProvides authentication, user management, realtor profiles, GoHighLevel CRM integration,\nonboarding webhooks, and administrative functions.\n\n## Authentication\nMost endpoints require a Bearer token obtained via the Login or Register endpoints.\nPass the token in the `Authorization` header as `Bearer <token>`.\n\n## Webhooks\nGoHighLevel webhook endpoints are public (CSRF exempt) and authenticated via\nthe `X-OmniReferral-Webhook` header secret.",
        "contact": {
            "email": "admin@omnireferrals.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://omnireferrals.com",
            "description": "OmniReferral Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/login": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Show login form",
                "operationId": "749bdaf6374ca2f6fdb149d27225370e",
                "responses": {
                    "200": {
                        "description": "Login page"
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and start session",
                "description": "Logs in a user with workspace role, email, and password. Starts a web session.",
                "operationId": "ec9c379d2a959de66404708dd4a9322d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to dashboard on success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/webhooks/gohighlevel/purchase": {
            "post": {
                "tags": [
                    "Webhooks",
                    "GoHighLevel"
                ],
                "summary": "Handle package purchase webhook from GoHighLevel",
                "description": "Receives a package purchase notification from GHL, creates/updates the user and realtor profile, dispatches sync and setup email.",
                "operationId": "2e8e2b0cd328d1a26de0b851b7e94393",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseWebhookPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase processed successfully"
                    },
                    "401": {
                        "description": "Unauthorized - invalid webhook secret"
                    },
                    "422": {
                        "description": "Missing email address"
                    },
                    "500": {
                        "description": "Account provisioning failed"
                    }
                }
            }
        },
        "/webhooks/gohighlevel/onboarding": {
            "post": {
                "tags": [
                    "Webhooks",
                    "GoHighLevel"
                ],
                "summary": "Receive onboarding completion from GoHighLevel",
                "description": "Handles onboarding form submission webhook. Creates/updates user, realtor/buyer profile, generates password token, and queues setup email.",
                "operationId": "9df21519ef544aec9be428c240823a74",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OnboardingWebhookPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Onboarding processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "dashboard": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid webhook secret"
                    },
                    "422": {
                        "description": "Missing email"
                    },
                    "500": {
                        "description": "Account provisioning failed"
                    }
                }
            }
        },
        "/webhooks/gohighlevel/lead-status": {
            "post": {
                "tags": [
                    "Webhooks",
                    "GoHighLevel"
                ],
                "summary": "Receive lead status update from GoHighLevel",
                "description": "Updates lead status (contacted, closed, etc.) and notifies the customer if needed.",
                "operationId": "f2f807182482f992df85f438fcdd6b1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lead_number": {
                                        "type": "string"
                                    },
                                    "ghl_contact_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "contacted",
                                            "qualified",
                                            "closed",
                                            "lost"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lead status synced"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lead not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "required": [
                    "role",
                    "email",
                    "password"
                ],
                "properties": {
                    "role": {
                        "description": "Workspace role",
                        "type": "string",
                        "enum": [
                            "agent",
                            "admin",
                            "staff"
                        ]
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    },
                    "remember": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "OnboardingWebhookPayload": {
                "required": [
                    "email"
                ],
                "properties": {
                    "contact_id": {
                        "description": "GoHighLevel contact ID",
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "name": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "agent",
                            "realtor",
                            "buyer",
                            "seller"
                        ]
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zip_code": {
                        "type": "string"
                    },
                    "brokerage_name": {
                        "type": "string"
                    },
                    "license_number": {
                        "type": "string"
                    },
                    "specialties": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "form_id": {
                        "type": "string"
                    },
                    "form_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PurchaseWebhookPayload": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "package_slug": {
                        "type": "string"
                    },
                    "package_id": {
                        "type": "integer"
                    },
                    "contact_id": {
                        "type": "string"
                    },
                    "brokerage_name": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "sanctum",
                "description": "Enter Sanctum token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}